requirements.txt
fastapi
uvicorn
chromadb
openai
python-dotenv
pydantic
pypdf
onnxruntime
cryptography